+++ lib/grell/capybara_driver.rb
@@ -14,7 +14,9 @@ module Grell
     def setup_capybara
       @poltergeist_driver = nil
 
-      driver_name = "poltergeist_crawler_#{Time.now.strftime('%m_%d_%y_%H_%M')}".to_sym
+      # Capybara will not re-run the block if the driver name already exists, so the driver name
+      # will have a time integer appended to ensure uniqueness.
+      driver_name = "poltergeist_crawler_#{Time.now.to_i}".to_sym
       Capybara.register_driver driver_name do |app|
         @poltergeist_driver = Capybara::Poltergeist::Driver.new(app, {
           js_errors: false,
@@ -32,7 +34,7 @@ module Grell
         "User-Agent" => USER_AGENT
       }
 
-      fail "Poltergeist Driver could not be properly initialized" unless @poltergeist_driver
+      raise "Poltergeist Driver could not be properly initialized" unless @poltergeist_driver
 
       @poltergeist_driver
     end