+++ lib/grell/crawler.rb
@@ -51,12 +51,21 @@ module Grell
       Grell.logger.info "Visiting #{site.url}, visited_links: #{@collection.visited_pages.size}, discovered #{@collection.discovered_pages.size}"
       site.navigate
       filter!(site.links)
+      add_redirect_url(site)
 
-      if block #The user of this block can send us a :retry to retry accessing the page
-        while block.call(site) == :retry
-          Grell.logger.info "Retrying our visit to #{site.url}"
-          site.navigate
-          filter!(site.links)
+      if block # The user of this block can send us a :retry to retry accessing the page
+        begin
+          while block.call(site) == :retry
+            Grell.logger.info "Retrying our visit to #{site.url}"
+            site.navigate
+            filter!(site.links)
+            add_redirect_url(site)
+          end
+        rescue Capybara::Poltergeist::BrowserError, Capybara::Poltergeist::DeadClient,
+               Capybara::Poltergeist::JavascriptError, Capybara::Poltergeist::StatusFailError,
+               Capybara::Poltergeist::TimeoutError, Errno::ECONNRESET, Timeout::Error, URI::InvalidURIError => e
+          site.unavailable_page(404, e)
+          return
         end
       end
 
@@ -77,6 +86,13 @@ module Grell
     def default_add_match
       Proc.new do |collection_page, page|
         collection_page.url.downcase == page.url.downcase
+      end
+    end
+
+    # Store the resulting redirected URL along with the original URL
+    def add_redirect_url(site)
+      if site.url != site.current_url
+        @collection.create_page(site.current_url, site.id)
       end
     end